Catch conflicts when writing

+++ src/backend/s3.ts
@@ -1,5 +1,11 @@
-import { Store, Etag, ListEntry } from "./store";
+import { Store, Etag, ListEntry, EtagMismatchError } from "./store";
 import { AwsClient } from "aws4fetch";
+
+export class HTTPError extends Error {
+  constructor(r: Response) {
+    super(`${r.status} ${r.statusText}`);
+  }
+}
 
 export class S3Store implements Store {
   client: AwsClient;
@@ -49,7 +55,7 @@ export class S3Store implements Store {
         prefix: this.prefix
       })
     );
-    if (!r.ok) throw r;
+    if (!r.ok) throw new HTTPError(r);
     const doc = (new DOMParser).parseFromString(await r.text(), "text/xml");
     for(let contents of doc.querySelectorAll("ListBucketResult > Contents")) {
       const key = this.childByTag(contents, "Key")?.textContent;
@@ -74,19 +80,36 @@ export class S3Store implements Store {
     }
     return out;
   }
+
+  async etag(path: string): Promise<Etag> {
+    const r: Response = await this.client.fetch(
+      this.constructUrl(this.prefix + path),
+      {
+        method: "HEAD",
+      },
+    );
+    if (r.status === 404) return null;
+    if (!r.ok) throw new HTTPError(r);
+    return r.headers.get("ETag");
+  }
+
   async get(path: string) {
     const r: Response = await this.client.fetch(this.constructUrl(this.prefix + path));
     if (r.status === 404) return {
       content: "",
       etag: null,
     };
-    if (!r.ok) throw r;
+    if (!r.ok) throw new HTTPError(r);
     return {
       content: await r.text(),
       etag: r.headers.get("ETag"),
     }
   }
-  async put(path: string, content: string, _etag: Etag) {
+
+  async put(path: string, content: string, etag: Etag) {
+    // S3 doesn't have the necessary APIs to avoid TOCTOU :(
+    if (await this.etag(path) !== etag) throw new EtagMismatchError();
+
     const r: Response = await this.client.fetch(
       this.constructUrl(this.prefix + path),
       {
@@ -94,19 +117,23 @@ export class S3Store implements Store {
         body: content,
       },
     );
-    if (!r.ok) throw r;
+    if (!r.ok) throw new HTTPError(r);
     return {
       etag: r.headers.get("ETag"),
     };
   }
-  async delete(path: string, _etag: Etag) {
+
+  async delete(path: string, etag: Etag) {
+    // S3 doesn't have the necessary APIs to avoid TOCTOU :(
+    if (await this.etag(path) !== etag) throw new EtagMismatchError();
+
     const r: Response = await this.client.fetch(
       this.constructUrl(this.prefix + path),
       {
         method: "DELETE",
       },
     );
-    if (!r.ok) throw r;
+    if (!r.ok) throw new HTTPError(r);
     return {
       etag: null,
     };
